requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated_types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
audioop_lts==0.2.2
av==16.1.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset_normalizer==3.4.4
click==8.3.1
cloup==3.0.8
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
distro==1.9.0
django_cors_headers==4.9.0
django==6.0.2
djangorestframework==3.16.1
filetype==1.2.0
frozenlist==1.8.0
glcontext==3.0.0
google_auth==2.48.0
google_genai==1.64.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isosurfaces==0.1.2
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain_core==1.2.19
langchain_google_genai==4.2.1
langchain_openai==1.1.11
langchain==1.2.10
langgraph_checkpoint==4.0.0
langgraph_prebuilt==1.0.8
langgraph_sdk==0.3.8
langgraph==1.0.9
langsmith==0.7.6
manim==0.20.0
manimpango==0.6.1
mapbox_earcut==2.0.0
markdown_it_py==4.0.0
mdurl==0.1.2
moderngl_window==3.1.1
moderngl==5.12.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pillow==12.1.1
pip==25.3
propcache==0.4.1
pyasn1_modules==0.4.2
pyasn1==0.6.2
pycairo==1.29.0
pycparser==3.0
pydantic_core==2.41.5
pydantic==2.12.5
pydub==0.25.1
pyglet==2.1.13
pyglm==2.8.3
pygments==2.19.2
python_dotenv==1.2.1
pyyaml==6.0.3
regex==2026.2.28
requests_toolbelt==1.0.0
requests==2.32.5
rich==14.3.3
rsa==4.9.1
scipy==1.17.1
screeninfo==0.8.1
skia_pathops==0.9.2
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
srt==3.5.3
svgelements==1.9.6
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing_extensions==4.15.0
typing_inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
